backend/requirements.txt
# Backend API Requirements
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Deep Learning
tensorflow==2.15.0
# OR tensorflow-macos==2.15.0  # For Apple Silicon Macs

# Image Processing
Pillow==10.1.0

# Data Processing
numpy==1.26.2

# Python Standard Library (usually pre-installed)
python-multipart==0.0.6  # For file uploads
